Add Profile dropdown with i18n
