+++ src/components/navbar/LeftSide/LeftSide.tsx
@@ -1,4 +1,3 @@
-import React from "react";
 import style from "./navleftside.module.scss";
