+++ src/components/navbar/LeftSide/navleftside.module.scss
