+++ src/components/navbar/Navbar.tsx
@@ -1,7 +1,7 @@
 import React from "react";
-import LeftSide from "./LeftSide";
-import SearchBar from "./SearchBar";
-import RightSide from "./RightSide";
+import LeftSide from "./LeftSide/LeftSide.tsx";
+import SearchBar from "./SearchBar/SearchBar.tsx";
+import RightSide from "./RightSide/RightSide.tsx";
 import style from "./navbar.module.scss"
 
 const Navbar = () => {
