+++ src/components/navbar/RightSide/navrightside.module.scss
@@ -1,0 +1,73 @@
+@use "@/styles/nav.scss" as *;
+
+.rightSide {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  gap: 5px;
+
+  
+}
+
+.profileIcon{
+  position: relative;
+ 
+  i {
+    font-size: $navProfileIconSize;
+    color: $navProfileIconColor;
+     cursor: pointer;
+  }
+  &:hover .profileDropdown{
+    transition: $navProfileDropdownTransition;
+  }
+}
+
+.languageSelect {
+  font-size: 14px;
+  padding: 5px;
+  border: none;
+  border-radius: 5px;
+  color: $navLanguageSelectColor;
+  cursor: pointer;
+  outline: none;
+}
+
+.profileDropdown{
+  position: absolute;
+  right: 0;
+  top: 55px;
+  background: $navProfileDropdownBGColor;
+  width: 150px;
+  height: auto;
+  border-radius: $navProfileDropdownBorderRadius;
+  pointer-events: none;
+  transition: $navProfileDropdownTransition;
+  ul{
+  font-size: $navProfileDropdownFontSize;
+  list-style: none;
+  display: flex;
+  justify-content: flex-start;
+  align-items: flex-start;
+  gap: 10px;
+  flex-direction: column;
+  padding: 10px;
+  li{
+    background: white;
+    width: 100%;
+    border-radius: 2px;
+    padding: 5px 0;
+    text-align: center;
+    color: $navProfileDropdownColor;
+    font-weight: 500;
+    cursor: pointer;
+  
+    &:hover{
+      color: $navProfileDropdownHoverColor;
+    }
+ 
+    i{
+       font-size: $navProfileDropdownFontSize;
+    }
+  }
+  }
+}
