+++ src/components/navbar/SearchBar/navsearch.module.scss
