+++ src/components/navbar/navbar.module.scss
@@ -7,6 +7,7 @@
   gap: 10px;
   padding: 0 20px 10px 20px;
   border-bottom: $navBorderBottomColor;
+  font-family: $navFontFamily;
 }
